requirements.txt
opencv-python==4.2.0
pytest==6.2.1
email_validator==1.1.1
urllib3==1.25.9
tcp-latency==0.0.10
pyspeedtest==1.2.7
speedtest-cli==2.1.3
pillow==7.2.0
img2pdf==0.4.0
sendgrid==6.7.0